Add MethodicalMachine discovery logic.
automat._discover.findMachines takes an FQPN that refers to a package,
module, class, or MethodicalMachine instance, and recursively yields
out machines and their FQPNS, where the machines are:

1) at module level; or
2) an instance on a class (including nested classes), and in the
   module where that class was defined.

The goal of 2) is to show machines only where they are defined, not
where they are used.

This adds a dependency on twisted.python.modules.