Stop using graphviz's private FORMATS set.
The argument parser uses it to fail early when given an unsupported
image format.  Since the graphviz library itself determines what image
formats it supports, the contents of the set isn't private, but its
location inside the graphviz Python package is.

This commit copies the set into _visualize to break the dependency on
this private API.